refactor(artifact): unify the anchor/selection model in one file (phase 4)
The comment anchor model was split across two files: types.ts held the
live Sel/Selection/ComposerState, while lib/layout.ts held ElementWire,
ParsedAnchor, and parseAnchor mixed in with the pure margin-placement
algorithm. Consolidate the whole anchor model (live + parsed) into types.ts
(which already carried selLabel), and leave lib/layout.ts as pure
arrangement — layoutPins, clamp, groupThreads. Move the parseAnchor tests
alongside into types.test.ts.

Type-only; no behavior change. Typecheck + full gate + unit tests green.
