+++ apps/web/src/pages/artifact/index.tsx
@@ -23,14 +23,14 @@ import { ActionsCtx } from "./comment-actions"
 import { CursorButton } from "./cursors/cursor-button"
 import { FloatingControl } from "./floating-control"
 import { canCommentWithRole, shouldPromptSignInToComment } from "./lib/comment-access"
-import { groupThreads, parseAnchor } from "./lib/layout"
+import { groupThreads } from "./lib/layout"
 import { parseRef, refFor } from "./parse-ref"
 import { PasswordGate } from "./password-gate"
 import { PublicViewer } from "./public-viewer"
 import { Presence } from "./rail-deck"
 import { ReviewCard } from "./review-card"
 import { SourceEditor } from "./source-editor"
-import type { ComposerState, PinItem } from "./types"
+import { type ComposerState, type PinItem, parseAnchor } from "./types"
 import { useArtifactFrame } from "./use-artifact-frame"
 import { useArtifactLive } from "./use-artifact-live"
 import { useCommentsPanel } from "./use-comments-panel"
